pyproject.toml
[tool.poetry]
name = "catch-them-all"
version = "0.1.0"
description = "Django application to catch them all"
authors = ["kenji <kenjitsuchiya1@gmail.com>"]
readme = "README.md"
packages = [{include = "catch_them_all"}]

[tool.poetry.dependencies]
python = "^3.11"
flake8 = "^6.0.0"
django = "^4.2.3"
requests = "^2.31.0"
psycopg2 = "^2.9.6"
psycopg = "^3.1.9"
pytest-django = "^4.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
